migrate throw ex to result for mediar

+++ src/Application/Features/Roles/Commands/Delete/DeleteRoleHandler.cs
@@ -1,5 +1,6 @@
-using Application.Common.Exceptions;
+using Application.Common.Errors;
 using Application.Common.Interfaces.Services.Identity;
+using Contracts.ApiWrapper;
 using Domain.Aggregates.Roles;
 using Mediator;
 using SharedKernel.Common.Messages;
@@ -7,21 +8,27 @@ using SharedKernel.Common.Messages;
 namespace Application.Features.Roles.Commands.Delete;
 
 public class DeleteRoleHandler(IRoleManagerService roleManagerService)
-    : IRequestHandler<DeleteRoleCommand>
+    : IRequestHandler<DeleteRoleCommand, Result<string>>
 {
-    public async ValueTask<Unit> Handle(
+    public async ValueTask<Result<string>> Handle(
         DeleteRoleCommand command,
         CancellationToken cancellationToken
     )
     {
-        Role role =
-            await roleManagerService.FindByIdAsync(command.RoleId)
-            ?? throw new NotFoundException(
-                [Messager.Create<Role>().Message(MessageType.Found).Negative().BuildMessage()]
+        Role? role = await roleManagerService.FindByIdAsync(command.RoleId);
+
+        if (role == null)
+        {
+            return Result<string>.Failure(
+                new NotFoundError(
+                    "Your Resource is not found",
+                    Messager.Create<Role>().Message(MessageType.Found).Negative().BuildMessage()
+                )
             );
+        }
 
         await roleManagerService.DeleteAsync(role);
 
-        return Unit.Value;
+        return Result<string>.Success();
     }
 }
